requirements.txt
apex
mmdet
mmcv-full
pycocotools; platform_system == "Linux"
torch
numpy